Back-reference from Assets & SaveData to Managers
In order for assets and save data to self-resolve references, they need a way to find their managers. While this creates a two-way dependency, this shouldn't be an issue with proper code; just don't add a single SaveData or Asset to two different managers. To prevent this from happening, a few checks have been added to throw if two managers are detected at once. Additionally, a bug was fixed that allowed writing disabled SaveDataManagers to delete files on the device.

+++ PMDDesktop.Server.Saving/SaveDataManager.cs
@@ -308,6 +308,9 @@ public class SaveDataManager : ISaveDataIndexable
 
 		string path = GetFilePath(data);
 
+		if (!WritingEnabled)
+			return;
+
 		// There's a chance we haven't had the chance to flush/save yet, so the file may not exist.
 		if (File.Exists(path))
 			File.Delete(path);
@@ -321,6 +324,14 @@ public class SaveDataManager : ISaveDataIndexable
 	public void Add(SaveData data)
 	{
 
+		if (data.Manager != null)
+		{
+			if (data.Manager == this)
+				throw new InvalidOperationException($"{data} is already assigned to {data.Manager}, the same manager it's trying to be added to.");
+			else
+				throw new InvalidOperationException($"{data} already has Manager {data.Manager} assigned to it.");
+		}
+
 		if (saveDatas.TryGetValue(data.GUID, out SaveData? blockingData))
 		{
 
@@ -333,6 +344,7 @@ public class SaveDataManager : ISaveDataIndexable
 
 		data.Dirty = true;
 		saveDatas.Add(data.GUID, data);
+		data.Manager = this;
 
 		Console.WriteLine($"Added SaveData ({data.GetType().Name}): {data}");
 
@@ -353,6 +365,7 @@ public class SaveDataManager : ISaveDataIndexable
 			{
 				saveDatas.Remove(data.GUID);
 				deleteQueue.Enqueue(data);
+				data.Manager = null;
 				return;
 			}
 			else
